+++ src/assets/styles/cmps/_User.scss
@@ -1,0 +1,78 @@
+.user-profile {
+	position: relative;
+
+	.user {
+		justify-content: space-between;
+		align-items: center;
+		width: 174px;
+		background-color: #0a0a0a;
+		border-radius: 20px;
+		transition: all 200ms ease-in-out;
+		margin-right: 30px;
+		margin-top: 10px;
+
+		@include for-mobile-layout {
+			width: 34vw;
+		}
+
+		&:hover {
+			cursor: pointer;
+			background-color: rgba($color: #ffffff59, $alpha: 0.3);
+			border: 1px solid black;
+		}
+
+		.img-warpper {
+			height: 33px;
+			img {
+				border: 1px #535353 solid;
+				width: 30px;
+				right: 10px;
+				border-radius: 50%;
+				transition: all 300ms ease-in;
+				background-color: #535353;
+				transition: all 300ms ease-in;
+				margin-left: 4px;
+				margin-top: 2px;
+			}
+		}
+
+		h4 {
+			font-size: 17px;
+			font-weight: 200;
+			transition: font-size 300ms ease-in;
+			margin: 6px 0 0.3em;
+		}
+
+		.svg-down-container {
+			margin-top: 4px;
+			margin-right: 5px;
+		}
+
+		svg {
+			fill: gainsboro;
+		}
+	}
+
+	.user-options-container {
+		background-color: #282828;
+		position: absolute;
+		min-width: 100px;
+		padding: 5px;
+		list-style-type: none;
+		margin: 10px 0;
+		cursor: pointer;
+
+		& > * {
+			padding: 5px;
+		}
+
+		& > *:not(:last-child) {
+			border-bottom: 1px solid #eeeeee29;
+		}
+
+		& > *:hover {
+			background-color: hsla(0, 0%, 100%, 0.219);
+			border-radius: 4px;
+		}
+	}
+}
