+++ src/assets/styles/main.scss
@@ -23,4 +23,5 @@
 @import 'cmps/AppHeader';
 @import 'cmps/AppFooter';
 @import 'cmps/UserMsg';
-@import 'cmps/NavBar'
+@import 'cmps/NavBar';
+@import 'cmps/User';
